Rethinking the activation sequence...
