+++ src/components/Annotation/ReplyForm/MobileFallback.css
@@ -16,6 +16,8 @@
 }
 
 .mobile-reply-form textarea {
+  border: none;
+  border-radius: 0;
   width: 100%;
   height: 100%;
 }
