+++ src/components/RichTextEditor/RichTextEditor.tsx
@@ -1,5 +1,8 @@
 import { RefObject, useMemo, useRef, useState } from 'react';
 import ReactQuill, { Range } from 'react-quill';
+import type { DeltaStatic } from 'quill';
+import { UrlDialog } from './UrlDialog';
+import type { Translations } from 'src/Types';
 import {
   ArrowCounterClockwise,
   ArrowClockwise,
@@ -7,18 +10,18 @@ import {
   Link,
   YoutubeLogo,
 } from '@phosphor-icons/react';
-import './styles.css';
-import type { DeltaStatic } from 'quill';
-import { UrlDialog } from './UrlDialog';
-import type { Translations } from 'src/Types';
+
+import './RichTextEditor.css';
 
 export interface RichTextEditorProps {
-  initialValue?: string;
-  value: string | DeltaStatic | undefined;
-  onChange(value: DeltaStatic): void;
-  placeholder?: string;
   editable?: boolean;
   i18n: Translations;
+  initialValue?: string;
+  placeholder?: string;
+  value: string | DeltaStatic | undefined;
+  onBlur(): void;
+  onChange(value: DeltaStatic): void;
+  onFocus(): void;
 }
 
 const CustomToolbar = () => (
@@ -198,14 +201,15 @@ export const RichTextEditor = (props: RichTextEditorProps) => {
     <>
       <div>
         {props.editable && <CustomToolbar />}
-        <ReactQuill
+        <ReactQuill          
+          ref={quill as RefObject<ReactQuill>}
           readOnly={!props.editable}
           value={props.value}
           modules={modules}
           formats={formats}
-          onChange={handleChange}
-          ref={quill as RefObject<ReactQuill>}
-        />
+          onChange={handleChange} 
+          onBlur={props.onBlur}
+          onFocus={props.onFocus} />
       </div>
       <UrlDialog
         title={title}
